moves osiris backend tests in seperate folders
