+++ tests/backends/osiris/test_GridFile.py
@@ -4,15 +4,10 @@ import numpy as np
 import pytest
 
 from nata.backends.osiris.hdf5 import Osiris_Dev_Hdf5_GridFile
-from nata.backends.osiris.hdf5 import Osiris_Dev_Hdf5_ParticleFile
 from nata.backends.osiris.hdf5 import Osiris_Hdf5_GridFile
-from nata.backends.osiris.hdf5 import Osiris_Hdf5_ParticleFile
 from nata.backends.osiris.zdf import Osiris_zdf_GridFile
-from nata.backends.osiris.zdf import Osiris_zdf_ParticleFile
 from nata.containers import GridDataset
-from nata.containers import ParticleDataset
 from nata.types import GridBackendType
-from nata.types import ParticleBackendType
 
 
 def test_Osiris_Hdf5_GridFile_isinstance_GridBackendType():
@@ -33,29 +28,6 @@ def test_GridDatasets_backends_are_registered():
     assert backends[Osiris_Hdf5_GridFile.name] is Osiris_Hdf5_GridFile
     assert backends[Osiris_Dev_Hdf5_GridFile.name] is Osiris_Dev_Hdf5_GridFile
     assert backends[Osiris_zdf_GridFile.name] is Osiris_zdf_GridFile
-
-
-def test_Osiris_Hdf5_ParticleFile_ParticleBackendType():
-    assert isinstance(Osiris_Hdf5_ParticleFile, ParticleBackendType)
-
-
-def test_Osiris_Dev_Hdf5_ParticleFile_ParticleBackendType():
-    assert isinstance(Osiris_Dev_Hdf5_ParticleFile, ParticleBackendType)
-
-
-def test_Osiris_zdf_ParticleFile_ParticleBackendType():
-    assert isinstance(Osiris_zdf_ParticleFile, ParticleBackendType)
-
-
-def test_ParticleDatasets_backends_are_registered():
-    backends = ParticleDataset.get_backends()
-
-    assert backends[Osiris_Hdf5_ParticleFile.name] is Osiris_Hdf5_ParticleFile
-    assert (
-        backends[Osiris_Dev_Hdf5_ParticleFile.name]
-        is Osiris_Dev_Hdf5_ParticleFile
-    )
-    assert backends[Osiris_zdf_ParticleFile.name] is Osiris_zdf_ParticleFile
 
 
 @pytest.fixture(name="os_hdf5_grid_444_file", scope="session")
